refactor: unify command usecase into one class
